refactor: 세션 패키지 분리
